chore: enum, struct 파일로 만들어 정리
